+++ CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - GPU execution: move the RNG state to CPU in case the checkpoint has been loaded on the GPU (#244)
 
+### Added
+
+- feat: add default batching to `predict` (#229)
+
 ## [0.23.0](https://github.com/owkin/connectlib/releases/tag/0.22.0) - 2022-08-01
 
 - feat: add compute task inputs (#236)